kwi: dezain o tyose simasita
